add bully leaderboard and refactor bully code to its own file
