+++ index.js
@@ -9,6 +9,7 @@ const client = new Client({ intents: [GatewayIntentBits.Guilds,
 require('dotenv').config();
 const storage = require('node-persist');
 const temperature = require('./temperature.js');
+const bully = require('./bully.js');
 
 const REACTION_CHANNEL_ID = process.env.REACTION_CHANNEL_ID;
 const PRONOUN_REACTION_POST_ID = process.env.PRONOUN_REACTION_POST_ID;
@@ -71,13 +72,16 @@ client.on("messageCreate", async msg => {
 
     if(isBullyCommand(command)){
         // switch doesn't deal well with wildcard matching so first check for a *ully command
-        await bullyHasHappened(msg, command);
+        await bully.bullyHasHappened(msg, command);
         return;
     }
 
     switch(command){
         case "!help":
             postHelp(msg);
+            break;
+        case "!bullyleaderboard":
+            bully.getLeaderboard(msg);
             break;
         case "!addcommand":
             if(userIsMod(msg)) await addCommand(msg);
@@ -102,7 +106,8 @@ client.on("messageCreate", async msg => {
  */
 function postHelp(msg){
     let response = '`!help` - display this message\n' +
-        '`!bully` - use this any time brandtamos is bullied\n';
+        '`!bully` - use this any time brandtamos is bullied\n' +
+        '`!bullyleaderboard` - show the current bullying leaderboard\n';
 
     //populate the rest of the help list from stored commands
     commandList.forEach((command) => {
@@ -112,90 +117,9 @@ function postHelp(msg){
     msg.channel.send(response);
 }
 
-/**
- * Calculates time since last bully event and updates timestamp
- * @async
- * @returns {Promise<number>} Time difference in milliseconds
- */
-async function getTimeSinceLastBully() {
-    let lastBullyTime = await storage.getItem("lastBullyTime");
-    const currentTimestamp = new Date().getTime();
-
-    // Handle NaN bullys (e.g., first run).
-    if (isNaN(lastBullyTime)) {
-        lastBullyTime = currentTimestamp;
-    }
-
-    await storage.setItem("lastBullyTime", currentTimestamp);
-
-    return currentTimestamp - lastBullyTime;
-}
-
-/**
- * Increments the bully counter and returns the new count
- * @async
- * @returns {Promise<number>} New bully count
- */
-async function incrementAndGetBullyCount() {
-    let bullyCount = await storage.getItem("bullyCount");
-
-    // Handle NaN bullys (e.g., first run).
-    if (isNaN(bullyCount)) {
-        bullyCount = 0;
-    }
-
-    const newBullyCount = bullyCount + 1;
-    await storage.setItem("bullyCount", newBullyCount);
-
-    return newBullyCount;
-}
-
-/**
- * Updates bully record if current time exceeds previous record
- * @async
- * @param {number} diffTime - Time difference in milliseconds
- * @returns {Promise<number>} Current record time in milliseconds
- */
-async function getAndSetBullyRecord(diffTime) {
-    let bullyRecord = await storage.getItem("bullyRecord");
-
-    // Handle NaN bullys (e.g., first run).
-    if (isNaN(bullyRecord)) {
-        bullyRecord = 0;
-    }
-
-    if (diffTime > bullyRecord) {
-        await storage.setItem("bullyRecord", diffTime);
-        return diffTime;
-    }
-    return bullyRecord;
-}
-
-/**
- * Handles bully tracking commands with dynamic output based on command used
- * @async
- * @param {import('discord.js').Message} msg - Discord message object
- * @param {string} command - The bully command used (e.g., !bully, !wully)
- */
-async function bullyHasHappened(msg, command){
-    const diffTime = await getTimeSinceLastBully();
-    const newBullyCount = await incrementAndGetBullyCount();
-    const currentBullyRecord = await getAndSetBullyRecord(diffTime);
-
-    const dayDiff = Math.floor((diffTime) / (1000 * 3600 * 24));
-    const hours = Math.floor((diffTime % (24 * 60 * 60 * 1000)) / (60 * 60 * 1000));
-    const minutes = Math.floor((diffTime % (60 * 60 * 1000)) / (60 * 1000));
-    const seconds = Math.floor((diffTime % (60 * 1000)) / 1000);
-
-    const recordDays = Math.floor((currentBullyRecord) / (1000 * 3600 * 24));
-    const recordHours = Math.floor((currentBullyRecord % (24 * 60 * 60 * 1000)) / (60 * 60 * 1000));
-    const recordMinutes = Math.floor((currentBullyRecord % (60 * 60 * 1000)) / (60 * 1000));
-    const recordSeconds = Math.floor((currentBullyRecord % (60 * 1000)) / 1000);
-
-    const ch = command.charAt(1).toLowerCase();
-    const bullyMsg = `${ch.toUpperCase()}enevolent moderator **${ch}randtamos** has ${ch}een ${ch}ullied.\n\nIt has ${ch}een **${dayDiff} days, ${hours} hours, ${minutes} minutes and ${seconds} seconds** since **${ch}randtamos** was last ${ch}ullied.\n\n**${ch}randtamos** has ${ch}een ${ch}ullied a total of **${newBullyCount} times**.\n\nThe record for the longest amount of time **${ch}randtamos** has not ${ch}een ${ch}ullied is **${recordDays} days, ${recordHours} hours, ${recordMinutes} minutes and ${recordSeconds} seconds**.`
-    msg.channel.send(bullyMsg);
-}
+/*async function handleBully(msg, command){
+    await bully.bullyHasHappened(msg, command);
+}*/
 
 /**
  * Adds a new custom command with description and output text
